ACMS-626: Refactor SiteConfigureForm to have default with acquia CMS
