+++ src/Form/SiteConfigureForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\acquia_cms_site_studio\Form;
+namespace Drupal\acquia_cms\Form;
 
 use Drupal\acquia_cms_tour\Form\AcquiaGoogleMapsAPIForm;
 use Drupal\Core\Config\ConfigFactoryInterface;
@@ -12,16 +12,9 @@ use Drupal\Core\Installer\Form\SiteConfigureForm as CoreSiteConfigureForm;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Extends the installer's site configuration form to configure Cohesion.
+ * Extends the installer's site configuration form for acquia CMS.
  */
-final class SiteConfigureForm extends ConfigFormBase {
-
-  /**
-   * The Cohesion API URL.
-   *
-   * @var string
-   */
-  private $apiUrl;
+class SiteConfigureForm extends ConfigFormBase {
 
   /**
    * The module installer.
@@ -42,22 +35,20 @@ final class SiteConfigureForm extends ConfigFormBase {
    *
    * @var \Drupal\Core\Installer\Form\SiteConfigureForm
    */
-  private $siteForm;
+  protected $siteForm;
 
   /**
    * The decorated Google Maps configuration form object.
    *
    * @var \Drupal\acquia_cms_tour\Form\AcquiaGoogleMapsAPIForm
    */
-  private $mapsForm;
+  protected $mapsForm;
 
   /**
    * SiteConfigureForm constructor.
    *
    * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
    *   The config factory service.
-   * @param string $api_url
-   *   The Cohesion API URL.
    * @param \Drupal\Core\Extension\ModuleInstallerInterface $module_installer
    *   The module installer.
    * @param \Drupal\Core\Extension\ModuleHandlerInterface $module_handler
@@ -67,9 +58,8 @@ final class SiteConfigureForm extends ConfigFormBase {
    * @param \Drupal\acquia_cms_tour\Form\AcquiaGoogleMapsAPIForm $maps_form
    *   The decorated Google Maps configuration form object.
    */
-  public function __construct(ConfigFactoryInterface $config_factory, string $api_url, ModuleInstallerInterface $module_installer, ModuleHandlerInterface $module_handler, CoreSiteConfigureForm $site_form, AcquiaGoogleMapsAPIForm $maps_form) {
+  public function __construct(ConfigFactoryInterface $config_factory, ModuleInstallerInterface $module_installer, ModuleHandlerInterface $module_handler, CoreSiteConfigureForm $site_form, AcquiaGoogleMapsAPIForm $maps_form) {
     parent::__construct($config_factory);
-    $this->apiUrl = $api_url;
     $this->moduleInstaller = $module_installer;
     $this->moduleHandler = $module_handler;
     $this->siteForm = $site_form;
@@ -82,7 +72,6 @@ final class SiteConfigureForm extends ConfigFormBase {
   public static function create(ContainerInterface $container) {
     return new static(
       $container->get('config.factory'),
-      $container->get('cohesion.api.utils')->getAPIServerURL(),
       $container->get('module_installer'),
       $container->get('module_handler'),
       CoreSiteConfigureForm::create($container),
@@ -111,22 +100,6 @@ final class SiteConfigureForm extends ConfigFormBase {
     $form = $this->siteForm->buildForm($form, $form_state);
     // Set default value for site name.
     $form['site_information']['site_name']['#default_value'] = $this->t('Acquia CMS');
-    $form['cohesion'] = [
-      'api_key' => [
-        '#type' => 'textfield',
-        '#title' => $this->t('API key'),
-        '#default_value' => getenv('SITESTUDIO_API_KEY'),
-      ],
-      'organization_key' => [
-        '#type' => 'textfield',
-        '#title' => $this->t('Organization key'),
-        '#default_value' => getenv('SITESTUDIO_ORG_KEY'),
-      ],
-      '#type' => 'details',
-      '#title' => $this->t('Acquia Site Studio'),
-      '#description' => $this->t('Enter your API key and organization key to automatically set up Acquia Site Studio (note that this process can take a while). If you do not want to use Site Studio right now, leave these fields blank -- you can always set it up later.'),
-      '#tree' => TRUE,
-    ];
 
     $form = $this->mapsForm->buildForm($form, $form_state);
     unset(
@@ -170,23 +143,6 @@ final class SiteConfigureForm extends ConfigFormBase {
       $this->mapsForm->submitForm($form, $form_state);
     }
 
-    $api_key = $form_state->getValue(['cohesion', 'api_key']);
-    $org_key = $form_state->getValue(['cohesion', 'organization_key']);
-
-    if ($api_key && $org_key) {
-      // For reasons I can't fathom, not resetting the config factory causes
-      // a non-interactive install (i.e., drush site:install) to be unable to
-      // load the API and organization keys in acquia_cms_install_tasks(), which
-      // in turn results in Cohesion's stuff not getting imported. So, although
-      // this may look bizarre, leave it as-is.
-      $this->resetConfigFactory();
-
-      $this->config('cohesion.settings')
-        ->set('api_url', $this->apiUrl)
-        ->set('api_key', $api_key)
-        ->set('organization_key', $org_key)
-        ->save(TRUE);
-    }
     // Enable the Acquia Telemetry module if user opt's in.
     $acquia_telemetry_opt_in = $form_state->getValue('acquia_telemetry');
     if ($acquia_telemetry_opt_in) {
